Update licence
The licence file didn't actually contain a licence , this was causing github not to show the project as open source.  
This text is from the daedalus repo where this is working correctly.  
This is NOT a change to the intent of the licence per se , although its possible the licence might not have applied before having not been explicitly mentioned in this file.