default.nix: disable check phases by default
The test suites of all cardano projects take many minutes (more than 10). For
building a docker image of the Icarus importer they are not necessary and create
a major overhead for trying out Icarus.

Add a bit of logic to replace `forceDontCheck` by `globalDoCheck`, which is not
a semantic negation and enables/disables checks for more libraries.